+++ tests/run.py
@@ -163,7 +163,7 @@ class TestRunner:
     def __init__(self):
         self.tests = []
         self.load_tests('tests')
-        self.default_classes = ['Object']
+        self.default_classes = ['Object', 'Error']
         self.default_classes_list_str = ''
         for item in self.default_classes:
             self.default_classes_list_str += "'" + item + '\', '